feat: 상품 정보 제거 시 사용자의 권한을 체크하는 기능을 추가한다

+++ src/main/java/cart/persistence/repository/MemberCartRepository.java
@@ -1,6 +1,7 @@
 package cart.persistence.repository;
 
 import cart.exception.ErrorCode;
+import cart.exception.ForbiddenException;
 import cart.exception.GlobalException;
 import cart.persistence.dao.CartDao;
 import cart.persistence.dao.MemberDao;
@@ -10,6 +11,7 @@ import cart.persistence.entity.MemberEntity;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Objects;
 
 @Repository
 public class MemberCartRepository {
@@ -33,8 +35,11 @@ public class MemberCartRepository {
         return cartDao.getProductsByMemberId(memberEntity.getId());
     }
 
-    public int deleteByMemberEmail(final String memberEmail, final Long productId) {
+    public int deleteByMemberEmail(final Long targetMemberId, final String memberEmail, final Long productId) {
         final MemberEntity memberEntity = getMemberEntity(memberEmail);
+        if (!Objects.equals(memberEntity.getId(), targetMemberId)) {
+            throw new ForbiddenException();
+        }
         return cartDao.deleteByMemberId(memberEntity.getId(), productId);
     }
